Correctly update the movement speed according to the default values
The given constant values inside the GameParameters class could be movement
per ms. This means we have to scale it according to our delta value.

The delta is set the time ()ms since the last tick/update. This means if a
computer cannot keep up with our target frame rate, the updates per one second
will be lower, but our delta value will be higher as normal.

Example:
0.3 * 10ms = 3 pixel
0.3 * 20ms = 6 pixel lower framerate per second, but more movement per second
   to have the same speed for our users

This would remove the need to have constant magic value of 17F. That might
differ for each user.